Revert "moved state out of navbar and made contact us in hero  and get in touch button in footer to open menu"
This reverts commit 0151b6387af661c0aa9f2acf55e8d3bb8d45ac57.
