+++ src/components/footer.js
@@ -8,7 +8,7 @@ import stackedLogo from "../images/logo-stacked-white.svg"
 import Grid from "@material-ui/core/Grid";
 import Button from '@material-ui/core/Button';
 
-const Footer = ({handleClickOpen}) => {
+const Footer = ({ className }) => {
 
 
   const divStyle = {
@@ -45,7 +45,6 @@ const Footer = ({handleClickOpen}) => {
               className={styles.button}
               component="a"
               href="#"
-              onClick={handleClickOpen}
             >Get in touch</Button>
           </Container>
         </Grid>
